requirements.txt
fastapi
uvicorn[standard]
lxml>=5.2.0
xmlschema
xsdata
PyYAML
Jinja2
cryptography
SQLAlchemy>=2.0.30
alembic>=1.13.2
psycopg[binary]>=3.1.19
